+++ js/lib/math.js
@@ -6,11 +6,11 @@
  * It features real and complex numbers, units, matrices, a large set of
  * mathematical functions, and a flexible expression parser.
  *
- * @version 6.3.0
- * @date    2019-12-31
+ * @version 6.4.0
+ * @date    2020-01-06
  *
  * @license
- * Copyright (C) 2013-2019 Jos de Jong <wjosdejong@gmail.com>
+ * Copyright (C) 2013-2020 Jos de Jong <wjosdejong@gmail.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy
@@ -51187,7 +51187,7 @@ var intersectDocs = {
 var distanceDocs = {
   name: 'distance',
   category: 'Geometry',
-  syntax: ['distance([x1, y1], [x2, y2])', 'distance([[x1, y1], [x2, y2])'],
+  syntax: ['distance([x1, y1], [x2, y2])', 'distance([[x1, y1], [x2, y2]])'],
   description: 'Calculates the Euclidean distance between two points.',
   examples: ['distance([0,0], [4,4])', 'distance([[0,0], [4,4]])'],
   seealso: []
@@ -53575,7 +53575,7 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
 
   /**
     * Calculates:
-    *    The eucledian distance between two points in 2 and 3 dimensional spaces.
+    *    The eucledian distance between two points in N-dimensional spaces.
     *    Distance between point and a line in 2 and 3 dimensional spaces.
     *    Pairwise distance between a set of 2D or 3D points
     * NOTE:
@@ -53587,6 +53587,7 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
     *-   math.distance({pointOneX: 4, pointOneY: 5}, {pointTwoX: 2, pointTwoY: 7})
     *    math.distance([x1, y1, z1], [x2, y2, z2])
     *    math.distance({pointOneX: 4, pointOneY: 5, pointOneZ: 8}, {pointTwoX: 2, pointTwoY: 7, pointTwoZ: 9})
+    *    math.distance([x1, y1, ... , N1], [x2, y2, ... , N2])
     *    math.distance([[A], [B], [C]...])
     *    math.distance([x1, y1], [LinePtX1, LinePtY1], [LinePtX2, LinePtY2])
     *    math.distance({pointX: 1, pointY: 4}, {lineOnePtX: 6, lineOnePtY: 3}, {lineTwoPtX: 2, lineTwoPtY: 8})
@@ -53607,6 +53608,7 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
     *    math.distance(
     *     {pointOneX: 4, pointOneY: 5, pointOneZ: 8},
     *     {pointTwoX: 2, pointTwoY: 7, pointTwoZ: 9})    // Returns 3
+    *    math.distance([1, 0, 1, 0], [0, -1, 0, -1])     // Returns 2
     *    math.distance([[1, 2], [1, 2], [1, 3]])         // Returns [0, 1, 1]
     *    math.distance([[1,2,4], [1,2,6], [8,1,3]])      // Returns [2, 7.14142842854285, 7.681145747868608]
     *    math.distance([10, 10], [8, 1, 3])              // Returns 11.535230316796387
@@ -53699,28 +53701,17 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
         }
 
         return _distancePointLine3D(x[0], x[1], x[2], y[0], y[1], y[2], y[3], y[4], y[5]);
-      } else if (x.length === 2 && y.length === 2) {
-        // Point to Point 2D
-        if (!_2d(x)) {
-          throw new TypeError('Array with 2 numbers or BigNumbers expected for first argument');
+      } else if (x.length === y.length && x.length > 0) {
+        // Point to Point N-dimensions
+        if (!_containsOnlyNumbers(x)) {
+          throw new TypeError('All values of an array should be numbers or BigNumbers');
         }
 
-        if (!_2d(y)) {
-          throw new TypeError('Array with 2 numbers or BigNumbers expected for second argument');
+        if (!_containsOnlyNumbers(y)) {
+          throw new TypeError('All values of an array should be numbers or BigNumbers');
         }
 
-        return _distance2d(x[0], x[1], y[0], y[1]);
-      } else if (x.length === 3 && y.length === 3) {
-        // Point to Point 3D
-        if (!_3d(x)) {
-          throw new TypeError('Array with 3 numbers or BigNumbers expected for first argument');
-        }
-
-        if (!_3d(y)) {
-          throw new TypeError('Array with 3 numbers or BigNumbers expected for second argument');
-        }
-
-        return _distance3d(x[0], x[1], x[2], y[0], y[1], y[2]);
+        return _euclideanDistance(x, y);
       } else {
         throw new TypeError('Invalid Arguments: Try again');
       }
@@ -53766,7 +53757,7 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
         }
 
         if ('pointOneX' in x && 'pointOneY' in x && 'pointTwoX' in y && 'pointTwoY' in y) {
-          return _distance2d(x.pointOneX, x.pointOneY, y.pointTwoX, y.pointTwoY);
+          return _euclideanDistance([x.pointOneX, x.pointOneY], [y.pointTwoX, y.pointTwoY]);
         } else {
           throw new TypeError('Key names do not match');
         }
@@ -53781,7 +53772,7 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
         }
 
         if ('pointOneX' in x && 'pointOneY' in x && 'pointOneZ' in x && 'pointTwoX' in y && 'pointTwoY' in y && 'pointTwoZ' in y) {
-          return _distance3d(x.pointOneX, x.pointOneY, x.pointOneZ, y.pointTwoX, y.pointTwoY, y.pointTwoZ);
+          return _euclideanDistance([x.pointOneX, x.pointOneY, x.pointOneZ], [y.pointTwoX, y.pointTwoY, y.pointTwoZ]);
         } else {
           throw new TypeError('Key names do not match');
         }
@@ -53819,6 +53810,15 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
     }
 
     return _isNumber(a[0]) && _isNumber(a[1]) && _isNumber(a[2]);
+  }
+
+  function _containsOnlyNumbers(a) {
+    // checks if the number of arguments are correct in count and are valid (should be numbers)
+    if (!Array.isArray(a)) {
+      a = _objectToArray(a);
+    }
+
+    return a.every(_isNumber);
   }
 
   function _parametricLine(a) {
@@ -53874,31 +53874,35 @@ Object(factory["a" /* factory */])(distance_name, distance_dependencies, functio
     return divideScalar(num, den);
   }
 
-  function _distance2d(x1, y1, x2, y2) {
-    var yDiff = subtract(y2, y1);
-    var xDiff = subtract(x2, x1);
-    var radicant = addScalar(multiplyScalar(yDiff, yDiff), multiplyScalar(xDiff, xDiff));
-    return sqrt(radicant);
-  }
+  function _euclideanDistance(x, y) {
+    var vectorSize = x.length;
+    var result = 0;
+    var diff = 0;
 
-  function _distance3d(x1, y1, z1, x2, y2, z2) {
-    var zDiff = subtract(z2, z1);
-    var yDiff = subtract(y2, y1);
-    var xDiff = subtract(x2, x1);
-    var radicant = addScalar(addScalar(multiplyScalar(zDiff, zDiff), multiplyScalar(yDiff, yDiff)), multiplyScalar(xDiff, xDiff));
-    return sqrt(radicant);
+    for (var i = 0; i < vectorSize; i++) {
+      diff = subtract(x[i], y[i]);
+      result = addScalar(multiplyScalar(diff, diff), result);
+    }
+
+    return sqrt(result);
   }
 
   function _distancePairwise(a) {
     var result = [];
+    var pointA = [];
+    var pointB = [];
 
     for (var i = 0; i < a.length - 1; i++) {
       for (var j = i + 1; j < a.length; j++) {
         if (a[0].length === 2) {
-          result.push(_distance2d(a[i][0], a[i][1], a[j][0], a[j][1]));
+          pointA = [a[i][0], a[i][1]];
+          pointB = [a[j][0], a[j][1]];
         } else if (a[0].length === 3) {
-          result.push(_distance3d(a[i][0], a[i][1], a[i][2], a[j][0], a[j][1], a[j][2]));
+          pointA = [a[i][0], a[i][1], a[i][2]];
+          pointB = [a[j][0], a[j][1], a[j][2]];
         }
+
+        result.push(_euclideanDistance(pointA, pointB));
       }
     }
 
@@ -59425,7 +59429,7 @@ Object(factory["a" /* factory */])(reviver_name, reviver_dependencies, function 
   };
 });
 // CONCATENATED MODULE: ./src/version.js
-var version = '6.3.0'; // Note: This file is automatically generated when building math.js.
+var version = '6.4.0'; // Note: This file is automatically generated when building math.js.
 // Changes made in this file will be overwritten.
 // CONCATENATED MODULE: ./src/plain/number/constants.js
 var constants_pi = Math.PI;